Fix everything except text rendering.
