+++ Cargo.toml
@@ -30,7 +30,6 @@ num-traits  = "0.2"
 nalgebra    = "0.15"
 ncollide3d  = "0.16"
 image       = "0.19"
-pistoncore-input = "0.21"
 
 [target.'cfg(not(target_arch = "wasm32"))'.dependencies]
 freetype    = "0.3"
