+++ src/camera/arc_ball.rs
@@ -1,7 +1,8 @@
 use camera::Camera;
+use event::{Action, Key, MouseButton, WindowEvent};
 use na::{self, Isometry3, Matrix4, Perspective3, Point3, Vector2, Vector3};
 use std::f32;
-use window::{Action, Canvas, Key, MouseButton, WindowEvent};
+use window::Canvas;
 
 /// Arc-ball camera mode.
 ///
