+++ src/event/mod.rs
@@ -1,0 +1,5 @@
+pub use self::event_manager::{Event, EventManager, Events};
+pub use self::window_event::{Action, Key, MouseButton, WindowEvent};
+
+mod event_manager;
+mod window_event;
